+++ src/io/HighsIO.h
@@ -16,7 +16,6 @@
 
 #include <iostream>
 #include <vector>
-#include "lp_data/HConst.h"
 
 class HighsOptions;
 
@@ -25,6 +24,16 @@ class HighsOptions;
  */
 enum class HighsMessageType { INFO, WARNING, ERROR };
 const char* const HighsMessageTypeTag[] = {"INFO", "WARNING", "ERROR"};
+
+enum HighsPrintMessageLevel {
+  ML_MIN = 0,
+  ML_NONE = ML_MIN,
+  ML_VERBOSE = 1,
+  ML_DETAILED = 2,
+  ML_MINIMAL = 4,
+  ML_ALWAYS = ML_VERBOSE | ML_DETAILED | ML_MINIMAL,
+  ML_MAX = ML_ALWAYS
+};
 
 /**
  * @brief Used to direct printed output to FILE* output, according
