+++ src/lp_data/HConst.h
@@ -44,16 +44,6 @@ enum class HighsDebugStatus {
   WARNING
 };
 
-enum HighsPrintMessageLevel {
-  ML_MIN = 0,
-  ML_NONE = ML_MIN,
-  ML_VERBOSE = 1,
-  ML_DETAILED = 2,
-  ML_MINIMAL = 4,
-  ML_ALWAYS = ML_VERBOSE | ML_DETAILED | ML_MINIMAL,
-  ML_MAX = ML_ALWAYS
-};
-
 enum class HighsOptionType { BOOL = 0, INT, DOUBLE, STRING };
 
 enum class HighsInfoType { INT = 1, DOUBLE };
